requirements.txt
#  Copyright © 2024 te4gh0st (Vitaliy Timtsurak). All rights reserved.
#  Licensed under the Apache License, Version 2.0

aiogram==2.25.2
python-dotenv==1.0.1
requests==2.31.0
loguru==0.7.2
beautifulsoup4==4.12.3
pip==23.2.1
aiohttp==3.8.6
certifi==2024.2.2
Babel==2.9.1
setuptools==68.2.0
